requirements.txt
kafka-python==2.0.2
psycopg2==2.8.6
pyyaml==5.3.1
requests==2.24.0
schedule==0.6.0